% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifiedSelf-package.R
\docType{package}
\name{quantifiedSelf-package}
\alias{quantifiedSelf}
\alias{quantifiedSelf-package}
\title{quantifiedSelf: Insight Timer Log Visualizer}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stephen Holsenbeck \email{sholsen@alumni.emory.edu}

}
\keyword{internal}
